Change KeepAlive server behavior.
Connection: KeepAlive is server's choice so the server could close the
connection suddenly even if a client send 'Connection: KeepAlive'.
(Apache httpd has 'MaxKeepAliveRequests' configuration for 'The maximum
number of requests to allow during a persistent connection' and the
httpd closes the persistent connection if count > config) Let
KeepAliveServer emulate this behavior.

test_keepalive in test_em-http-request.rb does not follow this behavior
at this moment. It would be need em-synchrony or something.